// src/interface/gyftr.d.ts
export type GyftrBrands = {
  _id: string;
  brandProductCode: string;
  __v: number;
  createdAt: string;
  data: {
    BrandProductCode: string;
    BrandName: string;
    Brandtype: string;
    RedemptionType?: string;
    OnlineRedemptionUrl?: string;
    BrandImage?: string;
    denominationList: string;
    stockAvailable: string;
    Category: string;
    Descriptions?: string;
    tnc?: string;
    importantInstruction?: string;
    redeemSteps?: string;
    updated_at?: string;
  };
  updatedAt: string;
};